Updated thumbnail image code so larger images can be used
This is in response to Facebook being crap about `og:image` sizes. I have chosen to use larger
Flickr images for thumbnails so they will show up when posts are shared on Facebook, but I still
want them to be square thumbnails on the page so I have used `object-fit: cover` in the css. I
know this probably won't look right on IE and Edge browsers, but for now this is the solution that
works for me