Set max-width for header when displayed as sidebar
Within the media queries elements of the scss added max-width declarations
to each of the relevant queries so they sidebar does not end up bigger
than the designated background image, which is 250px wide.

Now I just need to remember how to update the themes gem file so that the
website builds can be rerun to include it!